requirements.txt
flask==2.2.2
redis==4.3.4
requests==2.28.1
Flask-Caching==2.0.1
python-dotenv==0.15.0
